+++ src/index.ts
@@ -18,7 +18,5 @@ export * from './workload';
 export * from './daemon-set';
 export * from './role';
 export * from './role-binding';
-export * from './cluster-role';
-export * from './cluster-role-binding';
 
 export * from './api-resource.generated';
